Move all Workflow/Task prep funcs into new module
