+++ matflow/models/task.py
@@ -1,259 +1,16 @@
-"""`matflow.models.task.py`"""
+"""`matflow.models.task.py`
 
-import copy
+Module containing the Task and TaskSchema classes.
+
+"""
+
 import re
-from pathlib import Path
-from warnings import warn
 from pprint import pprint
 
 import numpy as np
 
-from matflow import CONFIG, TASK_SCHEMAS
-from matflow.models import CommandGroup, Command
-from matflow.jsonable import to_jsonable
-from matflow.utils import tile, repeat, arange
-from matflow.errors import (
-    SequenceError,
-    IncompatibleSequence,
-    TaskSchemaError,
-    TaskParameterError
-)
-
-
-def get_schema_dict(name, method, software_instance=None):
-    """Get the schema associated with the method/implementation of this task."""
-
-    match_task_idx = None
-    match_method_idx = None
-    match_imp_idx = None
-
-    for task_ref_idx, task_ref in enumerate(TASK_SCHEMAS):
-
-        if task_ref['name'] == name:
-
-            match_task_idx = task_ref_idx
-            for met_idx, met in enumerate(task_ref['methods']):
-
-                if met['name'] == method:
-
-                    match_method_idx = met_idx
-                    implementations = met.get('implementations')
-                    if implementations:
-
-                        for imp_idx, imp in enumerate(implementations):
-
-                            if imp['name'] == software_instance['name']:
-                                match_imp_idx = imp_idx
-                                break
-                    break
-            break
-
-    if match_task_idx is None:
-        msg = (f'No matching task found with name: "{name}"')
-        raise ValueError(msg)
-
-    if match_method_idx is None:
-        msg = (f'No matching method found with name: "{method}"'
-               f' in task: "{name}""')
-        raise ValueError(msg)
-
-    task_ref = TASK_SCHEMAS[match_task_idx]
-    met_ref = task_ref['methods'][met_idx]
-    inputs = task_ref.get('inputs', []) + met_ref.get('inputs', [])
-    outputs = task_ref.get('outputs', []) + met_ref.get('outputs', [])
-
-    imp_ref = None
-    in_map = None
-    out_map = None
-    command_opt = None
-
-    if match_imp_idx is not None:
-        imp_ref = met_ref['implementations'][match_imp_idx]
-
-        inputs += imp_ref.get('inputs', [])
-        outputs += imp_ref.get('outputs', [])
-
-        in_map = imp_ref.get('input_map', [])
-        out_map = imp_ref.get('output_map', [])
-        command_opt = imp_ref.get('commands', [])
-
-    outputs = list(set(outputs))
-
-    if software_instance:
-        implementation = software_instance['name']
-        command_group = {
-            'commands': command_opt,
-            'env_pre': software_instance.get('env_pre'),
-            'env_post': software_instance.get('env_post'),
-        }
-    else:
-        implementation = None
-        command_group = None
-
-    schema_dict = {
-        'name': name,
-        'method': method,
-        'implementation': implementation,
-        'inputs': inputs,
-        'outputs': outputs,
-        'input_map': in_map,
-        'output_map': out_map,
-        'command_group': command_group,
-    }
-
-    return schema_dict
-
-
-def normalise_local_inputs(base=None, sequences=None):
-    'Validate and normalise sequences and task inputs for a given task.'
-
-    if base is None:
-        base = {}
-    if sequences is None:
-        sequences = []
-
-    if not isinstance(sequences, list):
-        raise SequenceError('`sequences` must be a list.')
-
-    nest_req = True if len(sequences) > 1 else False
-
-    req_seq_keys = ['name', 'vals']
-    allowed_seq_keys = req_seq_keys + ['nest_idx']
-
-    prev_num_vals = None
-    prev_nest = None
-    inputs_lst = []
-    for seq in sequences:
-
-        miss_keys = list(set(req_seq_keys) - set(seq.keys()))
-        if miss_keys:
-            miss_keys_fmt = ', '.join([f'"{i}"' for i in miss_keys])
-            msg = f'Missing keys from sequence definition: {miss_keys_fmt}.'
-            raise SequenceError(msg)
-
-        bad_keys = list(set(seq.keys()) - set(allowed_seq_keys))
-        if bad_keys:
-            bad_keys_fmt = ', '.join([f'"{i}"' for i in bad_keys])
-            raise SequenceError(f'Unknown keys from sequence definition: {bad_keys_fmt}.')
-
-        if not isinstance(seq['vals'], list):
-            raise SequenceError(f'Sequence "{seq["name"]}" `vals` must be a list.')
-
-        if 'nest_idx' in seq:
-            if not isinstance(seq['nest_idx'], int) or (seq['nest_idx'] < 0):
-                msg = (f'`nest_idx` must be a positive integer or zero for sequence '
-                       f'"{seq["name"]}"')
-                raise SequenceError(msg)
-
-        if nest_req:
-            if 'nest_idx' not in seq:
-                msg = f'`nest_idx` is required for sequence "{seq["name"]}".'
-                raise SequenceError(msg)
-        else:
-            if 'nest_idx' in seq:
-                warn(f'`nest_idx` is not required for sequence "{seq["name"]}.')
-            seq['nest_idx'] = 0  # set a default
-
-        nest = seq['nest_idx']
-
-        num_vals = len(seq['vals'])
-        if prev_num_vals and nest == prev_nest:
-            # For same nest_idx, sequences must have the same lengths:
-            if num_vals != prev_num_vals:
-                msg = (f'Sequence "{seq["name"]}" shares a `nest_idx` with another '
-                       f'sequence but has a different number of values.')
-                raise IncompatibleSequence(msg)
-
-        prev_num_vals = num_vals
-        prev_nest = nest
-        inputs_lst.append(copy.deepcopy(seq))
-
-    for in_name, in_val in base.items():
-        inputs_lst.append({
-            'name': in_name,
-            'nest_idx': -1,
-            'vals': [copy.deepcopy(in_val)],
-        })
-
-    inputs_lst.sort(key=lambda x: x['nest_idx'])
-
-    return inputs_lst
-
-
-def get_local_inputs(schema_inputs, base=None, num_repeats=1, sequences=None,
-                     nest=True, merge_priority=None, groups=None):
-
-    inputs_lst = normalise_local_inputs(base, sequences)
-
-    if inputs_lst:
-        lengths = [len(i['vals']) for i in inputs_lst]
-        total_len = len(inputs_lst[0]['vals'])
-        for idx, i in enumerate(inputs_lst[1:], 1):
-            if i['nest_idx'] > inputs_lst[idx - 1]['nest_idx']:
-                total_len *= len(i['vals'])
-
-        prev_reps = total_len
-        prev_tile = 1
-        prev_nest = None
-
-    else:
-        total_len = 1
-
-    local_ins = {'inputs': {}}
-
-    for idx, input_i in enumerate(inputs_lst):
-
-        if (prev_nest is None) or (input_i['nest_idx'] > prev_nest):
-            rep_i = int(prev_reps / lengths[idx])
-            tile_i = int(total_len / (lengths[idx] * rep_i))
-            prev_nest = input_i['nest_idx']
-        else:
-            rep_i = prev_reps
-            tile_i = prev_tile
-
-        vals_idx = tile(repeat(arange(lengths[idx]), rep_i), tile_i)
-        vals_idx = repeat(vals_idx, num_repeats)
-        repeats_idx = tile(arange(num_repeats), total_len)
-
-        local_ins['repeats_idx'] = repeats_idx
-        local_ins['inputs'].update({
-            input_i['name']: {
-                # 'nest_idx': input_i['nest_idx'],
-                'vals': input_i['vals'],
-                'vals_idx': vals_idx,
-            }
-        })
-        prev_reps = rep_i
-        prev_tile = tile_i
-
-    allowed_grp = list(schema_inputs.keys()) + ['repeats']
-    allowed_grp_fmt = ', '.join([f'"{i}"' for i in allowed_grp])
-
-    def_group = {'default': {'nest': nest, 'group_by': allowed_grp}}
-    if merge_priority:
-        def_group.update({'merge_priority': merge_priority})
-
-    user_groups = {}
-    for group_name, group in (groups or {}).items():
-
-        if 'group_by' not in group:
-            raise ValueError(f'Missing `group_by` key in group {group_name}.')
-
-        for param in group['group_by']:
-            if param not in allowed_grp:
-                msg = (f'Parameter "{param}" cannot be grouped, because it '
-                       f'has no specified values. Allowed group values are: '
-                       f'{allowed_grp_fmt}.')
-                raise ValueError(msg)
-
-        user_groups.update({f'user_group_{group_name}': group})
-
-    all_groups = {**def_group, **user_groups}
-
-    local_ins['groups'] = all_groups
-    local_ins['length'] = total_len
-
-    return local_ins
+from matflow.models import CommandGroup
+from matflow.errors import TaskSchemaError, TaskParameterError
 
 
 class TaskSchema(object):
@@ -295,7 +52,7 @@ class TaskSchema(object):
 
     @property
     def input_contexts(self):
-        return list(set([i.get('context', '') for i in self.inputs]))
+        return list(set([i.get('context', None) for i in self.inputs]))
 
     def _validate_inputs_outputs(self):
         'Basic checks on inputs and outputs.'
